lcddriver.py and i2c_lib.py merged into drivers/i2c_dev.py; i2c_dev.py will automatically detect bus number and an i2c device (lcd) address from parsing i2cdetect and if fails, uses the default 0x27 lcd addr; soft requires i2c-tools
