+++ drivers/__init__.py
@@ -1,0 +1,1 @@
+from i2c_dev import Lcd
